Big refactoring (#22529)
* Big refactoring

* Clear imports

* Fix

* Fix compilation

---------
