+++ OsmAnd-java/src/main/java/net/osmand/search/SearchUICore.java
@@ -1108,8 +1108,8 @@ public class SearchUICore {
 				break;
 			case OBF_RESOURCE:
 				// sort order: DETAILED, WIKIPEDIA, TRAVEL, BASEMAP
-				int ord1 = o1.getObfType().ordinal();
-				int ord2 = o2.getObfType().ordinal();
+				int ord1 = o1.getResourceType().ordinal();
+				int ord2 = o2.getResourceType().ordinal();
 				if (ord1 != ord2) {
 					return ord2 > ord1 ? -1 : 1;
 				}
