+++ OsmAnd/src/net/osmand/plus/mapcontextmenu/editors/FavoritePointEditor.java
@@ -7,13 +7,13 @@ import net.osmand.data.Amenity;
 import net.osmand.data.FavouritePoint;
 import net.osmand.data.LatLon;
 import net.osmand.plus.views.layers.PlaceDetailsObject;
+import net.osmand.search.FullAmenitySearch;
 import net.osmand.shared.gpx.primitives.WptPt;
 import net.osmand.osm.edit.Entity;
 import net.osmand.plus.activities.MapActivity;
 import net.osmand.plus.myplaces.favorites.FavoriteGroup;
 import net.osmand.plus.plugins.osmedit.data.OpenstreetmapPoint;
 import net.osmand.plus.render.RenderingIcons;
-import net.osmand.plus.views.layers.MapSelectionHelper;
 import net.osmand.util.Algorithms;
 
 public class FavoritePointEditor extends PointEditor {
@@ -72,7 +72,8 @@ public class FavoritePointEditor extends PointEditor {
 			amenity = detailsObject.getSyntheticAmenity();
 		} else if (object instanceof OpenstreetmapPoint point) {
 			Entity entity = point.getEntity();
-			amenity = MapSelectionHelper.findAmenity(app, latLon, entity.getId(), null, null);
+			FullAmenitySearch searcher = app.getResourceManager().getAmenitySearcher();
+			amenity = searcher.findAmenity(latLon, entity.getId(), null, null);
 		}
 		if (amenity != null) {
 			setAmenity(amenity);
