+++ OsmAnd/src/net/osmand/plus/poi/PoiUIFilterDataProvider.java
@@ -10,6 +10,7 @@ import net.osmand.data.QuadRect;
 import net.osmand.plus.OsmandApplication;
 import net.osmand.plus.exploreplaces.ExplorePlacesProvider;
 import net.osmand.plus.views.layers.POIMapLayer.PoiUIFilterResultMatcher;
+import net.osmand.search.FullAmenitySearch;
 import net.osmand.util.MapUtils;
 
 import java.util.ArrayList;
@@ -45,7 +46,8 @@ public class PoiUIFilterDataProvider {
             return searchWikiOnline(lat, lon, topLatitude, bottomLatitude, leftLongitude, rightLongitude,
                     filter.wrapResultMatcher(matcher));
         } else {
-            return app.getResourceManager().searchAmenities(filter, filter.additionalFilter, topLatitude, leftLongitude,
+            FullAmenitySearch fullAmenitySearch = app.getResourceManager().getAmenitySearcher();
+            return fullAmenitySearch.searchAmenities(filter, filter.additionalFilter, topLatitude, leftLongitude,
                     bottomLatitude, rightLongitude, zoom, true, filter.wrapResultMatcher(matcher));
         }
     }
