+++ OsmAnd/src/net/osmand/plus/resources/ResourceManager.java
@@ -1,8 +1,6 @@
 package net.osmand.plus.resources;
 
 
-import static net.osmand.CollatorStringMatcher.StringMatcherMode.CHECK_EQUALS_FROM_SPACE;
-import static net.osmand.CollatorStringMatcher.StringMatcherMode.MULTISEARCH;
 import static net.osmand.IndexConstants.*;
 import static net.osmand.plus.AppInitEvents.ASSETS_COPIED;
 import static net.osmand.plus.AppInitEvents.MAPS_INITIALIZED;
@@ -18,23 +16,17 @@ import androidx.annotation.Nullable;
 
 import com.google.gson.Gson;
 
-import net.osmand.CollatorStringMatcher;
 import net.osmand.GeoidAltitudeCorrection;
 import net.osmand.IProgress;
-import net.osmand.Location;
 import net.osmand.PlatformUtil;
 import net.osmand.ResultMatcher;
 import net.osmand.binary.BinaryMapDataObject;
 import net.osmand.binary.BinaryMapIndexReader;
-import net.osmand.binary.BinaryMapIndexReader.SearchPoiAdditionalFilter;
-import net.osmand.binary.BinaryMapIndexReader.SearchPoiTypeFilter;
-import net.osmand.binary.BinaryMapIndexReader.SearchRequest;
 import net.osmand.binary.BinaryMapPoiReaderAdapter.PoiSubType;
 import net.osmand.binary.CachedOsmandIndexes;
-import net.osmand.binary.ObfConstants;
 import net.osmand.data.Amenity;
+import net.osmand.data.BaseDetailsObject;
 import net.osmand.data.LatLon;
-import net.osmand.data.QuadRect;
 import net.osmand.data.RotatedTileBox;
 import net.osmand.data.TransportRoute;
 import net.osmand.data.TransportStop;
@@ -63,10 +55,14 @@ import net.osmand.plus.resources.CheckAssetsTask.CheckAssetsListener;
 import net.osmand.plus.resources.ReloadIndexesTask.ReloadIndexesListener;
 import net.osmand.plus.utils.AndroidUtils;
 import net.osmand.plus.utils.FileUtils;
+import net.osmand.plus.views.layers.ContextMenuLayer;
 import net.osmand.plus.views.layers.MapTileLayer;
+import net.osmand.plus.views.layers.PlaceDetailsObject;
 import net.osmand.plus.views.layers.base.OsmandMapLayer.DrawSettings;
 import net.osmand.plus.wikipedia.WikipediaPlugin;
 import net.osmand.router.TransportStopsRouteReader;
+import net.osmand.search.FullAmenitySearch;
+import net.osmand.search.core.AmenityIndexRepository;
 import net.osmand.util.Algorithms;
 import net.osmand.util.MapUtils;
 
@@ -87,8 +83,6 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-
-import gnu.trove.map.hash.TIntObjectHashMap;
 
 /**
  * Resource manager is responsible to work with all resources
@@ -147,8 +141,6 @@ public class ResourceManager {
 	protected final Map<String, BinaryMapReaderResource> fileReaders = new ConcurrentHashMap<>();
 
 	protected final Map<String, RegionAddressRepository> addressMap = new ConcurrentHashMap<>();
-	protected final Map<String, AmenityIndexRepository> amenityRepositories = new ConcurrentHashMap<>();
-	//	protected final Map<String, BinaryMapIndexReader> routingMapFiles = new ConcurrentHashMap<>();
 	protected final Map<String, BinaryMapReaderResource> transportRepositories = new ConcurrentHashMap<>();
 	protected final Map<String, AmenityIndexRepository> travelRepositories = new ConcurrentHashMap<>();
 	protected final Map<String, String> indexFileNames = new ConcurrentHashMap<>();
@@ -170,6 +162,8 @@ public class ResourceManager {
 
 	private boolean depthContours;
 	private boolean indexesLoadedOnStart;
+
+	private final FullAmenitySearch fullAmenitySearch;
 
 	public ResourceManager(@NonNull OsmandApplication app) {
 		this.app = app;
@@ -199,6 +193,10 @@ public class ResourceManager {
 		if (!path.exists()) {
 			path.mkdir();
 		}
+
+		fullAmenitySearch = new FullAmenitySearch(searchAmenitiesInProgress,
+				fileName -> app.getTravelRendererHelper().getFileVisibilityProperty(fileName).get(),
+				app.getLanguage());
 	}
 
 	public BitmapTilesCache getBitmapTilesCache() {
@@ -358,14 +356,6 @@ public class ResourceManager {
 				: 0;
 	}
 
-	public void clearTileForMap(String file, ITileSource map, int x, int y, int zoom,
-			long requestTimestamp) {
-		TilesCache<?> cache = getTilesCache(map);
-		if (cache != null) {
-			cache.getTileForMap(file, map, x, y, zoom, true, false, true, requestTimestamp);
-		}
-	}
-
 	private GeoidAltitudeCorrection geoidAltitudeCorrection;
 	private boolean searchAmenitiesInProgress;
 
@@ -511,20 +501,6 @@ public class ResourceManager {
 		return warnings;
 	}
 
-	private void copyPoiTypes(boolean overwrite) {
-		try {
-			File file = app.getAppPath(SETTINGS_DIR + "poi_types.xml");
-			boolean exists = file.exists();
-			if (!exists || overwrite) {
-				FileOutputStream fout = new FileOutputStream(file);
-				Algorithms.streamCopy(MapPoiTypes.class.getResourceAsStream("poi_types.xml"), fout);
-				fout.close();
-			}
-		} catch (Exception e) {
-			log.error(e.getMessage(), e);
-		}
-	}
-
 	private void renameRoadsFiles(ArrayList<File> files, File roadsPath) {
 		Iterator<File> it = files.iterator();
 		while (it.hasNext()) {
@@ -665,7 +641,7 @@ public class ResourceManager {
 					boolean isTravelObf = resource.getFileName().endsWith(BINARY_TRAVEL_GUIDE_MAP_INDEX_EXT);
 					if (mapReader.containsPoiData()) {
 						AmenityIndexRepositoryBinary amenityResource = new AmenityIndexRepositoryBinary(f, resource, app);
-						amenityRepositories.put(fileName, amenityResource);
+						fullAmenitySearch.addAmenityRepository(fileName, amenityResource);
 						if (isTravelObf) {
 							// reuse until new BinaryMapReaderResourceType.TRAVEL_GPX
 							travelRepositories.put(resource.getFileName(), amenityResource);
@@ -702,7 +678,7 @@ public class ResourceManager {
 			}
 		}
 		Map<PoiCategory, Map<String, PoiType>> toAddPoiTypes = new HashMap<>();
-		for (AmenityIndexRepository repo : amenityRepositories.values()) {
+		for (AmenityIndexRepository repo : fullAmenitySearch.getAmenityRepositories()) {
 			Map<String, List<String>> categories = ((AmenityIndexRepositoryBinary) repo).getDeltaPoiCategories();
 			if (!categories.isEmpty()) {
 				for (Map.Entry<String, List<String>> entry : categories.entrySet()) {
@@ -752,7 +728,7 @@ public class ResourceManager {
 	}
 
 	public List<AmenityIndexRepository> getTravelGpxRepositories() {
-		return getAmenityRepositories(true);
+		return fullAmenitySearch.getAmenityRepositories(true);
 	}
 
 	public List<AmenityIndexRepository> getWikivoyageRepositories() {
@@ -775,100 +751,7 @@ public class ResourceManager {
 
 	////////////////////////////////////////////// Working with amenities ////////////////////////////////////////////////
 	public List<AmenityIndexRepository> getAmenityRepositories() {
-		return getAmenityRepositories(true);
-	}
-
-	public List<AmenityIndexRepository> getAmenityRepositories(boolean includeTravel) {
-		List<String> fileNames = new ArrayList<>(amenityRepositories.keySet());
-		List<AmenityIndexRepository> baseMaps = new ArrayList<>();
-		List<AmenityIndexRepository> result = new ArrayList<>();
-
-		fileNames.sort(Algorithms.getStringVersionComparator());
-
-		for (String fileName : fileNames) {
-			if (fileName.endsWith(BINARY_TRAVEL_GUIDE_MAP_INDEX_EXT)) {
-				if (!includeTravel || !app.getTravelRendererHelper().getFileVisibilityProperty(fileName).get()) {
-					continue;
-				}
-			}
-			AmenityIndexRepository r = amenityRepositories.get(fileName);
-			if (r != null && r.isWorldMap()) {
-				baseMaps.add(r);
-			} else if (r != null) {
-				result.add(r);
-			}
-		}
-
-		result.addAll(baseMaps);
-		return result;
-	}
-
-	@NonNull
-	public List<Amenity> searchAmenities(SearchPoiTypeFilter filter, QuadRect rect,
-			boolean includeTravel) {
-		return searchAmenities(filter, null, rect.top, rect.left, rect.bottom, rect.right, -1, includeTravel, null);
-	}
-
-	@NonNull
-	public List<Amenity> searchAmenities(SearchPoiTypeFilter filter, double top,
-			double left, double bottom,
-			double right, int zoom, boolean includeTravel,
-			ResultMatcher<Amenity> matcher) {
-		return searchAmenities(filter, null, top, left, bottom, right, zoom, includeTravel, matcher);
-	}
-
-	@NonNull
-	public List<Amenity> searchAmenities(SearchPoiTypeFilter filter,
-			SearchPoiAdditionalFilter additionalFilter, double topLatitude,
-			double leftLongitude, double bottomLatitude,
-			double rightLongitude, int zoom, boolean includeTravel,
-			ResultMatcher<Amenity> matcher) {
-
-		Set<Long> openAmenities = new HashSet<>();
-		Set<Long> closedAmenities = new HashSet<>();
-		List<Amenity> actualAmenities = new ArrayList<>();
-
-		searchAmenitiesInProgress = true;
-		try {
-			boolean isEmpty = filter.isEmpty();
-			if (isEmpty && additionalFilter != null) {
-				filter = null;
-			}
-			if (!isEmpty || additionalFilter != null) {
-				int top31 = MapUtils.get31TileNumberY(topLatitude);
-				int left31 = MapUtils.get31TileNumberX(leftLongitude);
-				int bottom31 = MapUtils.get31TileNumberY(bottomLatitude);
-				int right31 = MapUtils.get31TileNumberX(rightLongitude);
-
-				List<AmenityIndexRepository> repos = getAmenityRepositories(includeTravel);
-
-				for (AmenityIndexRepository repo : repos) {
-					if (matcher != null && matcher.isCancelled()) {
-						searchAmenitiesInProgress = false;
-						break;
-					}
-					if (repo.checkContainsInt(top31, left31, bottom31, right31)) {
-						List<Amenity> foundAmenities = repo.searchAmenities(top31, left31, bottom31, right31,
-								zoom, filter, additionalFilter, matcher);
-						if (foundAmenities != null) {
-							for (Amenity amenity : foundAmenities) {
-								Long id = amenity.getId();
-								if (amenity.isClosed()) {
-									closedAmenities.add(id);
-								} else if (!closedAmenities.contains(id)) {
-									if (openAmenities.add(id)) {
-										actualAmenities.add(amenity);
-									}
-								}
-							}
-						}
-					}
-				}
-			}
-		} finally {
-			searchAmenitiesInProgress = false;
-		}
-		return actualAmenities;
+		return fullAmenitySearch.getAmenityRepositories(true);
 	}
 
 	@NonNull
@@ -885,242 +768,9 @@ public class ResourceManager {
 		return new ArrayList<>(poiSubTypes);
 	}
 
-	public List<Amenity> searchAmenitiesOnThePath(List<Location> locations, double radius,
-			SearchPoiTypeFilter filter,
-			ResultMatcher<Amenity> matcher) {
-		searchAmenitiesInProgress = true;
-		List<Amenity> amenities = new ArrayList<>();
-		try {
-			if (locations != null && locations.size() > 0) {
-				List<AmenityIndexRepository> repos = new ArrayList<>();
-				double topLatitude = locations.get(0).getLatitude();
-				double bottomLatitude = locations.get(0).getLatitude();
-				double leftLongitude = locations.get(0).getLongitude();
-				double rightLongitude = locations.get(0).getLongitude();
-				for (Location l : locations) {
-					topLatitude = Math.max(topLatitude, l.getLatitude());
-					bottomLatitude = Math.min(bottomLatitude, l.getLatitude());
-					leftLongitude = Math.min(leftLongitude, l.getLongitude());
-					rightLongitude = Math.max(rightLongitude, l.getLongitude());
-				}
-				if (!filter.isEmpty()) {
-					for (AmenityIndexRepository index : getAmenityRepositories()) {
-						if (index.checkContainsInt(
-								MapUtils.get31TileNumberY(topLatitude),
-								MapUtils.get31TileNumberX(leftLongitude),
-								MapUtils.get31TileNumberY(bottomLatitude),
-								MapUtils.get31TileNumberX(rightLongitude))) {
-							repos.add(index);
-						}
-					}
-					if (!repos.isEmpty()) {
-						for (AmenityIndexRepository r : repos) {
-							List<Amenity> res = r.searchAmenitiesOnThePath(locations, radius, filter, matcher);
-							if (res != null) {
-								amenities.addAll(res);
-							}
-						}
-					}
-				}
-			}
-		} finally {
-			searchAmenitiesInProgress = false;
-		}
-		return amenities;
-	}
-
-	public boolean containsAmenityRepositoryToSearch(boolean searchByName) {
-		for (AmenityIndexRepository index : getAmenityRepositories()) {
-			if (searchByName) {
-				if (index instanceof AmenityIndexRepositoryBinary) {
-					return true;
-				}
-			} else {
-				return true;
-			}
-		}
-		return false;
-	}
-
-
-	private List<Amenity> searchRouteByName(String multipleSearch, CollatorStringMatcher.StringMatcherMode mode, ResultMatcher<Amenity> matcher) {
-		List<Amenity> result = new ArrayList<>();
-		SearchRequest<Amenity> req = BinaryMapIndexReader.buildSearchPoiRequest(
-				0, 0, multipleSearch, 0, Integer.MAX_VALUE, 0, Integer.MAX_VALUE, matcher
-		);
-		req.setMatcherMode(mode);
-		for (AmenityIndexRepository index : getAmenityRepositories(false)) {
-			List<Amenity> amenities = index.searchPoiByName(req);
-			if (!Algorithms.isEmpty(amenities)) {
-				result.addAll(amenities);
-			}
-		}
-		return result;
-	}
-
-	public List<Amenity> searchRoutePartOf(String routeId) {
-		ResultMatcher<Amenity> matcher = new ResultMatcher<Amenity>() {
-			@Override
-			public boolean publish(Amenity amenity) {
-				String members = amenity.getAdditionalInfo(Amenity.ROUTE_MEMBERS_IDS);
-				if (members != null) {
-					HashSet<String> ids = new HashSet<>();
-					Collections.addAll(ids, members.split(" "));
-					return ids.contains(routeId);
-				}
-				return false;
-			}
-
-			@Override
-			public boolean isCancelled() {
-				return false;
-			}
-		};
-		return searchRouteByName(routeId, CHECK_EQUALS_FROM_SPACE, matcher);
-	}
-
-	public Map<String, List<Amenity>> searchRouteMembers(String multipleSearch) {
-		HashSet<String> ids = new HashSet<>();
-		Collections.addAll(ids, multipleSearch.split(" "));
-		ResultMatcher<Amenity> matcher = new ResultMatcher<Amenity>() {
-			@Override
-			public boolean publish(Amenity amenity) {
-				String routeId = amenity.getAdditionalInfo(Amenity.ROUTE_ID);
-				return routeId != null && ids.contains(routeId);
-			}
-
-			@Override
-			public boolean isCancelled() {
-				return false;
-			}
-		};
-
-		Map<String, List<Amenity>> map = new HashMap<>();
-		List<Amenity> result = searchRouteByName(multipleSearch, MULTISEARCH, matcher);
-		for (Amenity am : result) {
-			String routeId = am.getAdditionalInfo(Amenity.ROUTE_ID);
-			List<Amenity> amenities = map.computeIfAbsent(routeId, l -> new ArrayList<>());
-			amenities.add(am);
-		}
-		for (String id : ids) {
-			if (!map.containsKey(id)) {
-				map.put(id, null);
-			}
-		}
-		return map;
-	}
-
-	public List<Amenity> searchAmenitiesByName(String searchQuery,
-			double topLatitude, double leftLongitude, double bottomLatitude, double rightLongitude,
-			double lat, double lon, ResultMatcher<Amenity> matcher) {
-		List<Amenity> amenities = new ArrayList<>();
-		List<AmenityIndexRepositoryBinary> list = new ArrayList<>();
-		int left = MapUtils.get31TileNumberX(leftLongitude);
-		int top = MapUtils.get31TileNumberY(topLatitude);
-		int right = MapUtils.get31TileNumberX(rightLongitude);
-		int bottom = MapUtils.get31TileNumberY(bottomLatitude);
-		for (AmenityIndexRepository index : getAmenityRepositories(false)) {
-			if (matcher != null && matcher.isCancelled()) {
-				break;
-			}
-			if (index instanceof AmenityIndexRepositoryBinary) {
-				if (index.checkContainsInt(top, left, bottom, right)) {
-					if (index.checkContains(lat, lon)) {
-						list.add(0, (AmenityIndexRepositoryBinary) index);
-					} else {
-						list.add((AmenityIndexRepositoryBinary) index);
-					}
-
-				}
-			}
-		}
-
-		// Not using boundaries results in very slow initial search if user has many maps installed
-//		int left = 0;
-//		int top = 0;
-//		int right = Integer.MAX_VALUE;
-//		int bottom = Integer.MAX_VALUE;
-		for (AmenityIndexRepositoryBinary index : list) {
-			if (matcher != null && matcher.isCancelled()) {
-				break;
-			}
-			List<Amenity> result = index.searchAmenitiesByName(MapUtils.get31TileNumberX(lon), MapUtils.get31TileNumberY(lat),
-					left, top, right, bottom,
-					searchQuery, matcher);
-			amenities.addAll(result);
-		}
-
-		return amenities;
-	}
-
-
 	public AmenityIndexRepositoryBinary getAmenityRepositoryByFileName(String filename) {
-		return (AmenityIndexRepositoryBinary) amenityRepositories.get(filename);
+		return (AmenityIndexRepositoryBinary) fullAmenitySearch.getAmenityRepository(filename);
 	}
-
-	@NonNull
-	public List<BinaryMapDataObject> searchBinaryMapDataForAmenity(@NonNull Amenity amenity, int limit) {
-		String name = amenity.getName();
-		long osmId = ObfConstants.getOsmObjectId(amenity);
-		ResultMatcher<BinaryMapDataObject> matcher = new ResultMatcher<>() {
-			@Override
-			public boolean publish(BinaryMapDataObject object) {
-				long id = ObfConstants.getOsmObjectId(object);
-				if (osmId != -1) {
-					return osmId == id;
-				} else if (!Algorithms.isEmpty(name)) {
-					TIntObjectHashMap<String> names = object.getObjectNames();
-					return names != null && names.containsValue(name);
-				}
-				return false;
-			}
-
-			@Override
-			public boolean isCancelled() {
-				return false;
-			}
-		};
-		return searchBinaryMapDataObjects(amenity.getLocation(), matcher, limit);
-	}
-
-	@NonNull
-	public List<BinaryMapDataObject> searchBinaryMapDataObjects(@NonNull LatLon latLon,
-			@Nullable ResultMatcher<BinaryMapDataObject> matcher, int limit) {
-		List<BinaryMapDataObject> list = new ArrayList<>();
-
-		int y = MapUtils.get31TileNumberY(latLon.getLatitude());
-		int x = MapUtils.get31TileNumberX(latLon.getLongitude());
-
-		SearchRequest<BinaryMapDataObject> request = BinaryMapIndexReader.buildSearchRequest(x,
-				x + 1, y, y + 1, 15, null, new ResultMatcher<>() {
-					@Override
-					public boolean publish(BinaryMapDataObject object) {
-						if (matcher == null || matcher.publish(object)) {
-							list.add(object);
-							return true;
-						}
-						return false;
-					}
-
-					@Override
-					public boolean isCancelled() {
-						return matcher != null && matcher.isCancelled()
-								|| limit != -1 && list.size() == limit;
-					}
-				});
-
-		for (AmenityIndexRepository repository : getAmenityRepositories(false)) {
-			if (matcher != null && matcher.isCancelled()) {
-				break;
-			}
-			if (repository.isPoiSectionIntersects(request)) {
-				repository.searchMapIndex(request);
-			}
-		}
-		return list;
-	}
-
-	////////////////////////////////////////////// Working with address ///////////////////////////////////////////
 
 	public RegionAddressRepository getRegionRepository(String name) {
 		return addressMap.get(name);
@@ -1142,8 +792,6 @@ public class ResourceManager {
 		}
 		return res;
 	}
-
-	////////////////////////////////////////////// Working with transport ////////////////////////////////////////////////
 
 	private List<BinaryMapIndexReader> getTransportRepositories(double topLat, double leftLon,
 			double bottomLat, double rightLon) {
@@ -1217,7 +865,7 @@ public class ResourceManager {
 	////////////////////////////////////////////// Closing methods ////////////////////////////////////////////////
 
 	public void closeFile(String fileName) {
-		amenityRepositories.remove(fileName);
+		fullAmenitySearch.removeAmenityRepository(fileName);
 		addressMap.remove(fileName);
 		transportRepositories.remove(fileName);
 		indexFileNames.remove(fileName);
@@ -1251,7 +899,7 @@ public class ResourceManager {
 		transportRepositories.clear();
 		travelRepositories.clear();
 		addressMap.clear();
-		amenityRepositories.clear();
+		fullAmenitySearch.clearAmenityRepositories();
 		for (BinaryMapReaderResource res : fileReaders.values()) {
 			res.close();
 		}
@@ -1445,5 +1093,54 @@ public class ResourceManager {
 		Algorithms.closeStream(is);
 
 		return modifiedTime != null && file.setLastModified(modifiedTime);
+	}
+
+	public FullAmenitySearch getAmenitySearcher() {
+		return fullAmenitySearch;
+	}
+
+	@Nullable
+	public Object fetchOtherData(@NonNull OsmandApplication app, @Nullable Object object) {
+		PlaceDetailsObject detailsObject = null;
+		long time = System.currentTimeMillis();
+		if (object instanceof Amenity amenity) {
+			ContextMenuLayer.IContextMenuProvider provider = app.getOsmandMap().getMapLayers().getPoiMapLayer();
+			LatLon latLon = amenity.getLocation();
+			BaseDetailsObject baseObject = fullAmenitySearch.findPlaceDetails(latLon, amenity.getOsmId(), null, amenity.getWikidata());
+			if (baseObject == null) {
+				detailsObject = new PlaceDetailsObject(amenity, provider, app.getLanguage());
+			} else {
+				detailsObject = new PlaceDetailsObject(baseObject, provider);
+			}
+		}
+
+		if (object instanceof PlaceDetailsObject) {
+			detailsObject = (PlaceDetailsObject) object;
+		}
+
+		if (detailsObject == null) {
+			return object;
+		}
+
+		if (!detailsObject.hasGeometry()) {
+			List<BinaryMapDataObject> dataObjects = fullAmenitySearch.searchBinaryMapDataForAmenity(detailsObject.getSyntheticAmenity(), 1);
+			for (BinaryMapDataObject dataObject : dataObjects) {
+				if (copyCoordinates(detailsObject, dataObject)) {
+					break;
+				}
+			}
+		}
+		log.debug("fetchOtherData time " + (System.currentTimeMillis() - time));
+		return detailsObject;
+	}
+
+	private static boolean copyCoordinates(@NonNull PlaceDetailsObject detailsObject,
+										   @NonNull BinaryMapDataObject mapObject) {
+		int pointsLength = mapObject.getPointsLength();
+		for (int i = 0; i < pointsLength; i++) {
+			detailsObject.addX(mapObject.getPoint31XTile(i));
+			detailsObject.addY(mapObject.getPoint31YTile(i));
+		}
+		return pointsLength > 0;
 	}
 }
