+++ OsmAnd/src/net/osmand/plus/search/QuickSearchHelper.java
@@ -16,6 +16,7 @@ import net.osmand.data.FavouritePoint;
 import net.osmand.data.LatLon;
 import net.osmand.data.PointDescription;
 import net.osmand.data.QuadRect;
+import net.osmand.search.FullAmenitySearch;
 import net.osmand.shared.gpx.primitives.WptPt;
 import net.osmand.map.WorldRegion;
 import net.osmand.osm.AbstractPoiType;
@@ -169,7 +170,8 @@ public class QuickSearchHelper implements ResourceListener {
 
 	public Amenity findAmenity(String name, double lat, double lon, String lang, boolean transliterate) {
 		QuadRect rect = MapUtils.calculateLatLonBbox(lat, lon, 15);
-		List<Amenity> amenities = app.getResourceManager().searchAmenities(ACCEPT_ALL_POI_TYPE_FILTER, rect, true);
+		FullAmenitySearch fullAmenitySearch = app.getResourceManager().getAmenitySearcher();
+		List<Amenity> amenities = fullAmenitySearch.searchAmenities(ACCEPT_ALL_POI_TYPE_FILTER, rect, true);
 
 		MapPoiTypes types = app.getPoiTypes();
 		for (Amenity amenity : amenities) {
