+++ OsmAnd/src/net/osmand/plus/shared/OsmAndContextImpl.kt
@@ -95,7 +95,7 @@ class OsmAndContextImpl(private val app: OsmandApplication) : OsmAndContext {
 	private fun searchNearestCity(latLon: KLatLon, callback: CityNameCallback) {
 		val cityTypes = City.CityType.entries.associateBy { it.name.lowercase() }
 		val rect = MapUtils.calculateLatLonBbox(latLon.latitude, latLon.longitude, CITY_SEARCH_RADIUS)
-		val cities = app.resourceManager.searchAmenities(object : SearchPoiTypeFilter {
+		val cities = app.resourceManager.amenitySearcher.searchAmenities(object : SearchPoiTypeFilter {
 			override fun accept(type: PoiCategory, subcategory: String): Boolean {
 				return cityTypes.containsKey(subcategory)
 			}
