+++ OsmAnd/src/net/osmand/plus/track/helpers/GpxReaderTask.java
@@ -24,6 +24,7 @@ import net.osmand.osm.PoiCategory;
 import net.osmand.plus.AppInitializeListener;
 import net.osmand.plus.AppInitializer;
 import net.osmand.plus.OsmandApplication;
+import net.osmand.search.FullAmenitySearch;
 import net.osmand.shared.gpx.GpxDbHelper;
 import net.osmand.shared.gpx.RouteActivityHelper;
 import net.osmand.plus.plugins.PluginsHelper;
@@ -190,7 +191,8 @@ class GpxReaderTask extends AsyncTask<Void, GpxDataItem, Void> {
 		for (City.CityType t : City.CityType.values()) {
 			cityTypes.put(t.name().toLowerCase(Locale.ROOT), t);
 		}
-		List<Amenity> cities = app.getResourceManager().searchAmenities(new SearchPoiTypeFilter() {
+		FullAmenitySearch searcher = app.getResourceManager().getAmenitySearcher();
+		List<Amenity> cities = searcher.searchAmenities(new SearchPoiTypeFilter() {
 			@Override
 			public boolean accept(PoiCategory type, String subcategory) {
 				return cityTypes.containsKey(subcategory);
