+++ OsmAnd/src/net/osmand/plus/views/layers/MapMarkersLayer.java
@@ -52,6 +52,7 @@ import net.osmand.plus.base.containers.ShiftedBitmap;
 import net.osmand.plus.mapmarkers.MapMarker;
 import net.osmand.plus.mapmarkers.MapMarkersHelper;
 import net.osmand.plus.render.OsmandDashPathEffect;
+import net.osmand.search.FullAmenitySearch;
 import net.osmand.shared.routing.ColoringType;
 import net.osmand.plus.settings.backend.OsmandSettings;
 import net.osmand.plus.utils.NativeUtilities;
@@ -717,7 +718,8 @@ public class MapMarkersLayer extends OsmandMapLayer implements IContextMenuProvi
 	public Amenity getMapObjectByMarker(@NonNull MapMarker marker) {
 		if (marker.mapObjectName != null && marker.point != null) {
 			String mapObjName = marker.mapObjectName.split("_")[0];
-			return MapSelectionHelper.findAmenity(getApplication(), marker.point, -1, Collections.singletonList(mapObjName), null);
+			FullAmenitySearch searcher = getApplication().getResourceManager().getAmenitySearcher();
+			return searcher.findAmenity(marker.point, -1, Collections.singletonList(mapObjName), null);
 		}
 		return null;
 	}
