+++ OsmAnd/src/net/osmand/plus/views/layers/PlaceDetailsObject.java
@@ -17,7 +17,15 @@ public class PlaceDetailsObject extends BaseDetailsObject {
 	private final List<SelectedMapObject> selectedObjects = new ArrayList<>();
 
 	public PlaceDetailsObject() {
-		super();
+		super("en");
+	}
+
+	public PlaceDetailsObject(BaseDetailsObject baseDetailsObject, @Nullable IContextMenuProvider provider) {
+		super(baseDetailsObject.getLang());
+		for (Object obj : baseDetailsObject.getObjects()) {
+			addObject(obj, provider);
+		}
+		combineData();
 	}
 
 	public PlaceDetailsObject(@NonNull Object object, @Nullable IContextMenuProvider provider, @Nullable String lang) {
