+++ OsmAnd/src/net/osmand/plus/wikivoyage/data/TravelObfGpxFileReader.java
@@ -34,7 +34,6 @@ import net.osmand.ResultMatcher;
 import net.osmand.binary.BinaryMapDataObject;
 import net.osmand.binary.BinaryMapIndexReader;
 import net.osmand.binary.BinaryMapIndexReader.SearchRequest;
-import net.osmand.binary.BinaryMapPoiReaderAdapter;
 import net.osmand.binary.HeightDataLoader;
 import net.osmand.data.Amenity;
 import net.osmand.data.QuadRect;
@@ -43,8 +42,8 @@ import net.osmand.osm.PoiCategory;
 import net.osmand.plus.Version;
 import net.osmand.plus.activities.MapActivity;
 import net.osmand.plus.base.BaseLoadAsyncTask;
-import net.osmand.plus.resources.AmenityIndexRepository;
 import net.osmand.plus.utils.FileUtils;
+import net.osmand.search.core.AmenityIndexRepository;
 import net.osmand.shared.gpx.GpxFile;
 import net.osmand.shared.gpx.GpxUtilities;
 import net.osmand.shared.gpx.RouteActivityHelper;
